Extract goal presentation components from PresentingYourGoal
